check for available events
