+++ server/pipeline/stepbuilder/stepBuilder.go
@@ -240,13 +240,16 @@ func (b *StepBuilder) environmentVariables(metadata metadata.Metadata, axis matr
 func (b *StepBuilder) toInternalRepresentation(parsed *yaml_types.Workflow, environ map[string]string, metadata metadata.Metadata, stepID int64) (*backend_types.Config, error) {
 	var secrets []compiler.Secret
 	for _, sec := range b.Secs {
-		if !sec.Match(b.Curr.Event) {
-			continue
+		events := []string{}
+		for _, event := range sec.Events {
+			events = append(events, string(event))
 		}
+
 		secrets = append(secrets, compiler.Secret{
 			Name:           sec.Name,
 			Value:          sec.Value,
 			AllowedPlugins: sec.Images,
+			Events:         events,
 		})
 	}
